be/requirements.txt
Flask==3.0.0
Flask-SocketIO==5.3.6
Flask-CORS==4.0.0
python-socketio==5.10.0
python-engineio==4.7.1
chromadb==0.4.18
groq==0.4.1
python-dotenv==1.0.0
SQLAlchemy==2.0.23
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9
requests==2.31.0
numpy>=1.21.0
sentence-transformers==2.2.2
langchain==0.1.0
langchain-community==0.0.10
